[Geomagic] Create class GeomagicVisualModel to handle the device display. This class is still handled by GeomagicDriver, it is only to move code in another class for more clarity
