+++ applications/plugins/Geomagic/CMakeLists.txt
@@ -14,10 +14,12 @@ set(GEOMAGIC_SRC_DIR src/Geomagic)
 set(HEADER_FILES
     ${GEOMAGIC_SRC_DIR}/config.h.in
     ${GEOMAGIC_SRC_DIR}/GeomagicDriver.h
+    ${GEOMAGIC_SRC_DIR}/GeomagicVisualModel.h
 )
 
 set(SOURCE_FILES	  
     ${GEOMAGIC_SRC_DIR}/GeomagicDriver.cpp
+    ${GEOMAGIC_SRC_DIR}/GeomagicVisualModel.cpp
     ${GEOMAGIC_SRC_DIR}/initPlugin.cpp
 )
